Minimally working version of wasi http call
Tested in jco only so far, need to figure out what data to throw over
the wasm boundary. I've gotten to the IncomingResponse struct in Rust,
and that struct is implemented in js, but the methods are not listed
in the ts type of the js version of IncominResponse, so presumably not
meant to be used in js.